fix warnings

+++ src/base/Classes.cpp
@@ -1214,7 +1214,7 @@ int64_t MilliSecondsBetween(const TDateTime & ANow, const TDateTime & AThen)
 {
   TDateTime Result;
   Result = floor(MilliSecondSpan(ANow, AThen));
-  return Result;
+  return static_cast<int64_t>(Result);
 }
 
 int64_t SecondsBetween(const TDateTime & ANow, const TDateTime & AThen)
